11/1 map implementations
I may have fixed it, I am too tired to tell, or check
